+++ CHANGELOG.md
@@ -3,13 +3,14 @@
 
 -------------------------------------------------------------------------------------------------------------
 
-# 5.6.4 (2021-04-19)
+# 5.6.4 (2021-04-20)
 
 ### 新特性
 * 【core   】     DatePattern补充DateTimeFormatter（pr#308@Gitee）
 * 【core   】     DateUtil.compare增加支持给定格式比较（pr#310@Gitee）
 * 【core   】     BeanUtil增加edit方法（issue#I3J6BG@Gitee）
 * 【db     】     Column中加入columnDef字段默认值（issue#I3J6BG@Gitee）
+* 【core   】     BeanUtil增加copyToList方法（issue#1526@Github）
 
 ### Bug修复
 * 【db     】     修复SQL分页时未使用别名导致的错误，同时count时取消order by子句（issue#I3IJ8X@Gitee）